+++ lib/middleman-aks/category_manager.rb
@@ -1,5 +1,6 @@
 require 'middleman-aks/processor'
 require 'middleman-blog/uri_templates'
+require 'middleman-blog/extension'
 
 module Middleman::Aks
   class CategoryManager < Processor
@@ -15,11 +16,16 @@ module Middleman::Aks
     end
     ################
     module Helpers
-      def link_to_category(category, caption=nil)
+=begin
+      def link_to_category(category, blog_name=nil)
+      #def category_path(category, blog_name = nil)
         caption ||= h(category)
-        link_to(caption, aks.processors.category_manager.url_for(category))
+        #link_to(caption, aks.processors.category_manager.url_for(category))
+        link_to(caption, "categories/#{category}.html")
+        #link_to(caption, blog_controller(blog_name).category_pages.link(category))
       end
-      end
+=end
+    end
     ################
     class << self
       def activate
@@ -34,25 +40,23 @@ module Middleman::Aks
     include Middleman::Blog::UriTemplates
     def initialize(app, controller, options = {})
       super
-      @template = controller.options.category_template
-      @uri_template = controller.options.category_uri_template
-      #@template = "proxy_templates/category_template.html"
-      #@uri_template = "categories/{category}.html"
-      app.ignore @template
+#      @template = controller.options.category_template
+#      @uri_template = uri_template(controller.options.categorylink)
+#      app.ignore @template
       
       self.class.activate
     end
+=begin
     def page_for(category)
       app.sitemap.find_resource_by_path(url_for(catgory))
     end
     def url_for(category)
-      apply_uri_template(uri_template(@uri_template), {category: category})
+      apply_uri_template(@uri_template, {category: category})
     end
     def manipulate_resource_list(resources)
       newres = app.blog.categories.map do |category, articles|
         path = url_for(category)
         Middleman::Sitemap::Resource.new(app.sitemap, path).tap {|r|
-          $stderr.puts "proxy for #{category} on #{path}: #{@template}"
           r.proxy_to(@template)
           data = {locals: {category: category, articles: articles}}
           r.add_metadata data
@@ -60,5 +64,6 @@ module Middleman::Aks
       end
       resources + newres
     end
+=end
   end ## CategoryManager
 end ## module Middleman::Aks