+++ lib/middleman-aks/extension.rb
@@ -1,5 +1,7 @@
-require 'middleman-aks/controller'
+#require 'middleman-aks/controller'
 require 'middleman-aks/helpers'
+
+require 'middleman-aks/category_manager'
 
 module Middleman
   module Aks
@@ -16,58 +18,30 @@ module Middleman
     class Extension < Middleman::Extension
       helpers do
         include Middleman::Aks::Helpers
-        # general utils
-        def page_for(path)
-          sitemap.find_resource_by_path(path)
-        end
-        alias_method :resource_for, :page_for
 
 =begin
-        def link_to_page(page)
-          link_to(h(page.title), page)
+        def aks
+          @aks
+        end
+        def aks=(ext)
+          @aks = ext
         end
 =end
-        def top_page
-          sitemap.find_resource_by_path("/#{index_file}")
-        end
-
-        # controller accessors
-        def aks_controller
-          @_aks_controller
-        end
-        
-        def aks_controller=(controller)
-          @_aks_controller = controller
-        end
-        alias_method :aks, :aks_controller
-        alias_method :aks=, :aks_controller=
       end ## helpers
 
       ################
       # option settings
       #
       option :category_template, "proxy_templates/category_template.html"
-      option :category_uri_template, "categories/{category}.html"
-      
-=begin
-      option :index_template, "/templates/index_template.html"
-      option :archives_template_year, "/templates/archives_template_year.html"
-      option :archives_template_month, "/templates/archives_template_month.html"
-      option :archives_path_year, "/archives/%{year}/index.html"
-      option :archives_path_month, "/archives/%{year}/%{month}.html"
-
-      option :numbering_headings, true, "show numbering for headers"
-=end
-      
+      option :categorylink, "categories/{category}.html"
+      attr_reader :processors
       def initialize(klass, options_hash={}, &block)
         super
-        klass.set :aks_settings, options
+        #klass.set :aks_settings, options
 
         Middleman::Blog::BlogArticle.class_eval do
           include Middleman::Aks::SummaryText::InstanceMethodsToBlogArticle
         end
-
-        
       end
       
       ################
@@ -75,8 +49,17 @@ module Middleman
       def after_configuration
         #app.logger.level = 0   ## yet: debug
         app.logger.debug "- extension: after_configuration"
+#        app.aks = self
 
-        app.aks = Middleman::Aks::Controller.new(app, self)
+        if prefix = app.blog_controller.options.prefix
+          options.categorylink = File.join(prefix, options.categorylink)
+        end
+        
+        require 'ostruct'
+        @processors =
+          OpenStruct.new(category_manager: CategoryManager.new(app, self))
+
+        #app.aks = Middleman::Aks::Controller.new(app, self)
       end
     end
   end